backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
alembic
pydantic
python-dotenv
pdfplumber
pandas
python-multipart
passlib[bcrypt]
python-jose[cryptography]
openai
requests
